Started an LLM Project for CalPoly LAEP that will use LLMS
This project is for a Spring quarter position at my university, CalPoly SLO. We will use different LLM's to aide students' senior projects.

I started this project by forking a ChatGPT clone. I had to set up a firebase authorization, OpenAI API key, and update the code to work with the updated features of the OpenAI API.

There are still bugs to fix on the front end with chat logs, messages loading format, and styling issues.